+++ src/V1/validators/index.ts
@@ -33,4 +33,3 @@
 //
 
 export * from "./mustBe";
-export * from "./validate";